[CSL-1279] Remove 'SscGStateClass', create Pos.Ssc.Mem
