+++ lib/src/Pos/Lrc/Worker.hs
@@ -53,8 +53,9 @@ import           Pos.Lrc.Logic                         (findAllRichmenMaybe)
 import           Pos.Lrc.Mode                          (LrcMode)
 import           Pos.Reporting                         (reportMisbehaviour)
 import           Pos.Slotting                          (MonadSlots)
-import           Pos.Ssc.Extra                         (MonadSscMem, sscCalculateSeed)
-import           Pos.Ssc.GodTossing                    (noReportNoSecretsForEpoch1)
+import           Pos.Ssc.GodTossing                    (MonadSscMem,
+                                                        noReportNoSecretsForEpoch1,
+                                                        sscCalculateSeed)
 import           Pos.Ssc.GodTossing.Network.Constraint (GtMessageConstraints)
 import           Pos.Update.DB                         (getCompetingBVStates)
 import           Pos.Update.Poll.Types                 (BlockVersionState (..))
