+++ lib/src/Pos/Ssc/GodTossing.hs
@@ -11,37 +11,43 @@ module Pos.Ssc.GodTossing
        ( module Pos.Arbitrary.Ssc.GodTossing
        , module Pos.Ssc.GodTossing.Configuration
        , module Pos.Ssc.Core
+       , module Pos.Ssc.Extra
        , module Pos.Ssc.SeedError
        , module Pos.Ssc.VerifyError
        , module Pos.Ssc.Listeners
        , module Pos.Ssc.Worker
        , module Pos.Ssc.Lrc
+       , module Pos.Ssc.Mem
        , module Pos.Ssc.GodTossing.Functions
        , module Pos.Ssc.LocalData
        , module Pos.Ssc.GodTossing.SecretStorage
        , module Pos.Ssc.GodTossing.Seed
-       , module Pos.Ssc.GodTossing.GState
+       , module Pos.Ssc.GState
        , module Pos.Ssc.GodTossing.Toss
        , module Pos.Ssc.Types
        , module Pos.Ssc.GodTossing.Types.Message
        , module Pos.Ssc.GodTossing.VssCertData
+       , module Pos.Ssc.Util
        ) where
 
 import           Pos.Arbitrary.Ssc.GodTossing
 import           Pos.Binary.Ssc                   ()
 import           Pos.Ssc.Core
+import           Pos.Ssc.Extra
 import           Pos.Ssc.GodTossing.Configuration
 import           Pos.Ssc.GodTossing.Functions
-import           Pos.Ssc.GodTossing.GState
 import           Pos.Ssc.GodTossing.SecretStorage
 import           Pos.Ssc.GodTossing.Seed
 import           Pos.Ssc.GodTossing.Toss
 import           Pos.Ssc.GodTossing.Types.Message
 import           Pos.Ssc.GodTossing.VssCertData
+import           Pos.Ssc.GState
 import           Pos.Ssc.Listeners
 import           Pos.Ssc.LocalData
 import           Pos.Ssc.Lrc
+import           Pos.Ssc.Mem
 import           Pos.Ssc.SeedError
 import           Pos.Ssc.Types
+import           Pos.Ssc.Util
 import           Pos.Ssc.VerifyError
 import           Pos.Ssc.Worker
