+++ lib/src/Test/Pos/Block/Logic/Mode.hs
@@ -103,9 +103,9 @@ import           Pos.Slotting                   (HasSlottingVar (..), MonadSlots
                                                  getCurrentSlotSimple,
                                                  mkSimpleSlottingVar)
 import           Pos.Slotting.MemState          (MonadSlotsData)
+import           Pos.Ssc.GodTossing             (HasGtConfiguration, SscMemTag, SscState,
+                                                 mkSscState)
 import           Pos.Ssc.Types                  (SscBlock)
-import           Pos.Ssc.Extra                  (SscMemTag, SscState, mkSscState)
-import           Pos.Ssc.GodTossing             (HasGtConfiguration)
 import           Pos.Txp                        (GenericTxpLocalData, MempoolExt,
                                                  MonadTxpLocal (..), TxpGlobalSettings,
                                                  TxpHolderTag, mkTxpLocalData,
