--- ssc/Pos/Ssc/Class.hs
@@ -1,8 +1,0 @@
--- | Re-exports of Pos.Ssc.Class.* modules.
-
-module Pos.Ssc.Class
-       ( module Pos.Ssc.Class.Storage
-       ) where
-
-
-import           Pos.Ssc.Class.Storage
