+++ wallet/test/Test/Pos/Wallet/Web/Mode.hs
@@ -77,9 +77,9 @@ import           Pos.Shutdown                      (HasShutdownContext (..),
                                                     ShutdownContext (..))
 import           Pos.Slotting                      (HasSlottingVar (..), MonadSlots (..),
                                                     MonadSlotsData)
-import           Pos.Ssc.Extra                     (SscMemTag, SscState)
 import           Pos.Ssc.GodTossing.Configuration  (HasGtConfiguration)
-import           Pos.Ssc.Types                     (SscBlock)
+import           Pos.Ssc.Mem                       (SscMemTag)
+import           Pos.Ssc.Types                     (SscBlock, SscState)
 import           Pos.StateLock                     (StateLock, StateLockMetrics (..),
                                                     newStateLock)
 import           Pos.Txp                           (GenericTxpLocalData, MempoolExt,
